move awaitable_wait_for_x to infra
