+++ silkworm/sentry/rlpx/auth/handshake.cpp
@@ -17,7 +17,7 @@
 #include "handshake.hpp"
 
 #include <silkworm/infra/common/log.hpp>
-#include <silkworm/sentry/common/awaitable_wait_for_one.hpp>
+#include <silkworm/infra/concurrency/awaitable_wait_for_one.hpp>
 #include <silkworm/sentry/common/timeout.hpp>
 #include <silkworm/sentry/rlpx/framing/framing_cipher.hpp>
 #include <silkworm/sentry/rlpx/rlpx_common/disconnect_message.hpp>
@@ -29,7 +29,7 @@
 namespace silkworm::sentry::rlpx::auth {
 
 using namespace std::chrono_literals;
-using namespace common::awaitable_wait_for_one;
+using namespace concurrency::awaitable_wait_for_one;
 using common::Message;
 
 boost::asio::awaitable<AuthKeys> Handshake::auth(common::SocketStream& stream) {
